requirements.txt
# Core
chromadb==0.5.5
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
rich>=13.7.0

# Embeddings & Reranking
sentence-transformers>=3.0.0
# Cross-encoder lives inside sentence-transformers; we import from cross_encoder
torch>=2.2.0

# (Optional) LLM client
openai>=1.40.0

# CLI & API (optional)
fastapi>=0.112.0
uvicorn>=0.30.0

# For auto-screenshots of outputs
Pillow>=10.2.0
